refactor : modify MPA to SPA
