+++ styles/Room.module.css
@@ -1,10 +1,8 @@
 .lobby_wrapper{
-    padding: 200px;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    min-height: 100vh;
 }
 
 .lobby_wrapper > h2{
