+++ styles/User.module.css
@@ -1,10 +1,8 @@
 .login_wrapper{
-    padding: 200px;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    min-height: 100vh;
 }
 
 .login_wrapper > h2{
